Dialog message
Remove hyperscript redirect
Add link between pages
Add schema definition for swagger
Split large files
