+++ app/database/mongo/versions.py
@@ -12,7 +12,8 @@ from app.app_exception import (IncorrectTicketCount, ProjectNotRegistered,
 from app.conf import mongo_string
 from app.database.mongo.db_settings import DashCollection
 from app.database.settings import registered_projects
-from app.schema.project_schema import Bugs, StatusEnum, UpdateTickets, UpdateVersion
+from app.schema.project_schema import Bugs, StatusEnum, UpdateVersion
+from app.schema.ticket_schema import UpdateTickets
 
 
 async def clean_update_version(body: UpdateVersion) -> dict:
