+++ app/templates/base_project.html
@@ -5,16 +5,24 @@
 <div id="project_{{project_name}}" class="container">
   <ul class="nav nav-tabs">
   <li class="nav-item">
-    <a class={% if versions %} "nav-link  active" aria-current="page" {%else%} "nav-link"{% endif %}  href="/front/v1/projects/{{project_name}}">Versions</a>
+    <a class={% if versions %} "nav-link  active" aria-current="page" {%else%} "nav-link"{% endif %}
+       hx-get="/front/v1/projects/{{project_name}}"
+       hx-headers='{"eaid-request":"REDIRECT"}'>Versions</a>
   </li>
   <li class="nav-item">
-    <a class={% if campaigns %}"nav-link  active" aria-current="page"{%else %} "nav-link" {% endif %} href="/front/v1/projects/{{project_name}}/campaigns">Campaigns</a>
+    <a class={% if campaigns %}"nav-link  active" aria-current="page"{%else %} "nav-link" {% endif %}
+       hx-get="/front/v1/projects/{{project_name}}/campaigns"
+       hx-headers='{"eaid-request":"REDIRECT"}'>Campaigns</a>
   </li>
   <li class="nav-item">
-    <a class={% if repository %}"nav-link active" aria-current="page"{%else%} "nav-link" {% endif %}  href="/front/v1/projects/{{project_name}}/repository">Repository</a>
+    <a class={% if repository %}"nav-link active" aria-current="page"{%else%} "nav-link" {% endif %}
+        hx-get="/front/v1/projects/{{project_name}}/repository"
+        hx-headers='{"eaid-request":"REDIRECT"}'>Repository</a>
   </li>
     <li class="nav-item">
-    <a class={% if bugs %} "nav-link active" aria-current="page"{%else%} "nav-link" {% endif %}  href="/front/v1/projects/{{project_name}}/bugs">Bugs</a>
+    <a class={% if bugs %} "nav-link active" aria-current="page"{%else%} "nav-link" {% endif %}
+      hx-get="/front/v1/projects/{{project_name}}/bugs"
+      hx-headers='{"eaid-request":"REDIRECT"}'>Bugs</a>
   </li>
 </ul>
 {% block project %}
