+++ app/templates/tables/campaign_table.html
@@ -15,7 +15,9 @@
         <td>{{campaign.occurrence}}</td>
         <td>{{campaign.status}}</td>
           <td><button class="btn btn-primary"
-                      _="on click go to url /front/v1/projects/{{project_name}}/campaigns/{{campaign.version}}/{{campaign.occurrence}}">Manage</button> </td>
+                      hx-get="/front/v1/projects/{{project_name}}/campaigns/{{campaign.version}}/{{campaign.occurrence}}"
+                      hx-headers='{"eaid-request": "REDIRECT"}'
+                      >Manage</button> </td>
       </tr>
       {% endfor %}
       </tbody>
